shift allow only vertical move

+++ base/plugins/scenario/source/Process/Temporal/TemporalScenarioPresenter.cpp
@@ -651,7 +651,7 @@ void TemporalScenarioPresenter::on_eventCreated_impl(EventModel* event_model)
 	auto event_presenter = new EventPresenter{event_model,
 											  event_view,
 											  this};
-	event_view->setPos({rect.x() + event_model->date().msec() / m_millisecPerPixel,
+    event_view->setPos({qreal(rect.x() + event_model->date().msec() / m_millisecPerPixel),
 						rect.y() + rect.height() * event_model->heightPercentage()});
 
 	m_events.push_back(event_presenter);
@@ -735,7 +735,7 @@ void TemporalScenarioPresenter::on_constraintCreated_impl(TemporalConstraintView
 													constraint_view,
 													this};
 
-	constraint_view->setPos({rect.x() + constraint_view_model->model()->startDate().msec() / m_millisecPerPixel,
+    constraint_view->setPos({qreal(rect.x() + constraint_view_model->model()->startDate().msec() / m_millisecPerPixel),
                              rect.y() + rect.height() * constraint_view_model->model()->heightPercentage()});
 
 	constraint_presenter->on_horizontalZoomChanged(m_horizontalZoomSliderVal);
